% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast3d-package.R
\docType{package}
\name{cast3d-package}
\alias{cast3d}
\alias{cast3d-package}
\title{cast3d: Image to 3D Model}
\description{
A lightweight R client for the Tripo 3D generation API. Convert images or text prompts into 3D models (GLB format) and interactively view them in RStudio, RMarkdown, or Shiny via a custom Three.js widget. Built on httr2 for HTTP communication and htmlwidgets for 3D rendering.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yulab-github/cast3d}
  \item Report bugs at \url{https://github.com/yulab-github/cast3d/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

}
\keyword{internal}
