% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{glb_viewer-shiny}
\alias{glb_viewer-shiny}
\alias{glb_viewer_output}
\alias{render_glb_viewer}
\title{Shiny bindings for glb_viewer}
\usage{
glb_viewer_output(output_id, width = "100\%", height = "500px")

render_glb_viewer(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{output_id}{Output variable to read the value from.}

\item{width}{Widget width.}

\item{height}{Widget height.}

\item{expr}{An expression that generates a glb_viewer widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression?}
}
\description{
Output and render functions for using glb_viewer within Shiny
applications.
}
