% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripo3d-package.R
\docType{package}
\name{tripo3d-package}
\alias{tripo3d}
\alias{tripo3d-package}
\title{tripo3d: Image to 3D Model via Tripo API}
\description{
A lightweight R client for the Tripo 3D generation API. Convert images or text prompts into 3D models (GLB format) and interactively view them in RStudio, RMarkdown, or Shiny via a custom Three.js widget. Built on httr2 for HTTP communication and htmlwidgets for 3D rendering.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yulab-github/tripo3d}
  \item Report bugs at \url{https://github.com/yulab-github/tripo3d/issues}
}

}
\author{
\strong{Maintainer}: Yu Gongchao \email{yugongchao@example.com}

}
\keyword{internal}
